pyproject.toml
# pyproject.toml
[project]
name = "shadow-portfolio"
version = "0.1.0"
description = "Shadow Portfolio is a platform for creating and managing your portfolio of assets."
requires-python = ">=3.13"
authors = [
    { name="Ivan Rivera", email="ivan.rivera@shadowportfolio.com" }
]
dependencies = [
    "fastapi>=0.115.0",
    "requests>=2.28.0",
	"uvicorn>=0.40.0",
	"gunicorn>=25.0.3",
	"httpx>=0.28.1",
	"pydantic>=2.12.5",
	"pydantic-settings>=2.13.0",
	"alpaca-py>=0.43.2",
	"sqlmodel>=0.0.33",
	"finnhub-python>=2.4.27",
	"twelvedata>=1.2.28",
	"pysqlite3>=0.6.0",
	"braintrust>=0.5.5",
	"agno>=2.4.8",
	"mem0ai>=1.0.3",
	"pyyaml>=6.0.3",
	"jsonschema>=4.26.0",
]

[dependency-groups]
test = [
	"pytest>=9.0.2",
	"pytest-mock>=3.15.1",
	"hypothesis>=6.151.6",
	"pytest-cov>=7.0.0",
]

dev = [
	{ include-group = "test" },
	"ruff>=0.15.0",
	"pre-commit>=4.5.1",
	"mypy>=1.19.1",
	"commitizen>=4.13.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-preference = "only-managed"

[tool.hatch.build.targets.wheel]
packages = ["portfolio"]

[tool.pytest]
addopts = ["--cov=portfolio", "--cov-report=term-missing", "-ra", "-q", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
pythonpath = ["portfolio"]
markers = [
	"unit",
	"integration",
	"e2e",
	"slow",
]
filterwarnings = ["error", "ignore::DeprecationWarning", "ignore::FutureWarning", "ignore::ResourceWarning"]

[tool.coverage.run]
source = ["portfolio"]
branch = true
parallel = true
omit = ["tests/*", "*.__init__.py", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
	"if __name__ == '__main__':",
	"if TYPE_CHECKING:",
	"# pragma: no cover",
	"@abstractmethod",
	"raise NotImplementedError",
]
fail_under = 0

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["portfolio", "tests", "scripts"]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = [
	"E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "PTH", "RUF", "ASYNC", "S",
	"A", "DTZ", "T10", "ICN", "LOG", "G", "PIE", "PYI", "RSE", "RET", "SLF", "SLOT",
	"INT", "PGH", "PL", "TRY", "FLY", "PERF", "FURB", "T20", "FA", "ISC", "COM", "TID",
]
ignore = ["E501", "S101", "PLR2004", "TRY003", "PLR0913", "PGH003"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG"]

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 120

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
pretty = true
explicit_package_bases = true
namespace_packages = true
plugins = ["pydantic.mypy"]
disable_error_code = ["type-arg", "no-any-return"]

[[tool.mypy.overrides]]
module = ["braintrust.*", "finnhub.*", "twelvedata.*"]
ignore_missing_imports = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "uv"
major_version_zero = true
update_changelog_on_bump = true
tag_regex = "v\\d+\\.\\d+\\.\\d+"
annotate_tag = true
